# Performance improvements to MakeKey and HashKey APIs