```
name                                old time/op    new time/op    delta
EscapeMeasurement/this_is_a_test-8     142ns ± 3%      17ns ± 4%   -87.93%  (p=0.008 n=5+5)
EscapeMeasurement/this,is,a,test-8     195ns ± 1%     101ns ± 4%   -48.05%  (p=0.008 n=5+5)
Tags_HashKey/5_tags-no_esc-8          1.04µs ± 2%    0.25µs ±14%   -75.78%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-no_esc-8         4.85µs ± 2%    1.00µs ± 4%   -79.27%  (p=0.008 n=5+5)
Tags_HashKey/5_tags-esc-8             1.79µs ±11%    0.82µs ± 1%   -53.86%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-esc-8            8.02µs ± 0%    4.14µs ± 0%   -48.34%  (p=0.016 n=4+5)
MakeKey/this_is_a_test-8               327ns ± 1%      73ns ± 1%   -77.70%  (p=0.016 n=5+4)
MakeKey/this,is,a,test-8               392ns ± 3%     163ns ± 3%   -58.48%  (p=0.008 n=5+5)
MakeKey/this\_is\_a\_test-8            458ns ± 2%     293ns ± 3%   -35.97%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#01-8           2.04µs ± 1%    0.40µs ± 0%   -80.50%  (p=0.016 n=5+4)
MakeKey/this,is,a,test#01-8           2.12µs ± 1%    0.50µs ± 1%   -76.33%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#02-8           2.60µs ± 2%    0.94µs ± 2%   -63.75%  (p=0.008 n=5+5)
MakeKey/this,is,a,test#02-8           2.64µs ± 1%    1.03µs ± 3%   -60.84%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
EscapeMeasurement/this_is_a_test-8     32.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
EscapeMeasurement/this,is,a,test-8     62.0B ± 0%     32.0B ± 0%      ~     (p=0.079 n=4+5)
Tags_HashKey/5_tags-no_esc-8            368B ± 0%      128B ± 0%   -65.22%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-no_esc-8         1.86kB ± 0%    0.58kB ± 0%   -68.97%  (p=0.008 n=5+5)
Tags_HashKey/5_tags-esc-8               528B ± 0%      288B ± 0%   -45.45%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-esc-8            2.72kB ± 0%    2.72kB ± 0%      ~     (all equal)
MakeKey/this_is_a_test-8               64.0B ± 0%     16.0B ± 0%   -75.00%  (p=0.008 n=5+5)
MakeKey/this,is,a,test-8               93.4B ± 1%     64.0B ± 0%   -31.48%  (p=0.008 n=5+5)
MakeKey/this\_is\_a\_test-8            95.6B ± 1%     80.0B ± 0%   -16.32%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#01-8             832B ± 0%      224B ± 0%   -73.08%  (p=0.008 n=5+5)
MakeKey/this,is,a,test#01-8             878B ± 0%      272B ± 0%   -69.01%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#02-8             976B ± 0%      352B ± 0%   -63.93%  (p=0.008 n=5+5)
MakeKey/this,is,a,test#02-8           1.01kB ± 0%    0.42kB ± 0%   -58.65%  (p=0.008 n=5+5)

name                                old allocs/op  new allocs/op  delta
EscapeMeasurement/this_is_a_test-8      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
EscapeMeasurement/this,is,a,test-8      2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Tags_HashKey/5_tags-no_esc-8            2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-no_esc-8           2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Tags_HashKey/5_tags-esc-8               12.0 ± 0%      11.0 ± 0%    -8.33%  (p=0.008 n=5+5)
Tags_HashKey/25_tags-esc-8              52.0 ± 0%      52.0 ± 0%      ~     (all equal)
MakeKey/this_is_a_test-8                4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.008 n=5+5)
MakeKey/this,is,a,test-8                4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.008 n=5+5)
MakeKey/this\_is\_a\_test-8             4.00 ± 0%      3.00 ± 0%   -25.00%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#01-8             7.00 ± 0%      2.00 ± 0%   -71.43%  (p=0.008 n=5+5)
MakeKey/this,is,a,test#01-8             7.00 ± 0%      3.00 ± 0%   -57.14%  (p=0.008 n=5+5)
MakeKey/this_is_a_test#02-8             15.0 ± 0%      10.0 ± 0%   -33.33%  (p=0.008 n=5+5)
MakeKey/this,is,a,test#02-8             15.0 ± 0%      11.0 ± 0%   -26.67%  (p=0.008 n=5+5)
```